improve redability by separating modules in lang-doc
commit-id:a85fff87
