+++ crates/cairo-lang-doc/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod db;
 pub mod documentable_formatter;
 pub mod documentable_item;
+pub mod helpers;
 pub mod location_links;
 pub mod parser;
 pub mod signature_data;
